refactor: move misc. types to the types file
